Fix Windows file sharing. ListMmf doesn't allow file sharing.

+++ src/ListMmfTests/SmallestInt64Tests.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using BruSoftware.ListMmf;
 using FluentAssertions;
 using Xunit;
@@ -18,18 +17,12 @@ public class SmallestInt64Tests : IDisposable
     public SmallestInt64Tests()
     {
         s_count++;
-        if (File.Exists(TestPath))
-        {
-            File.Delete(TestPath);
-        }
+        if (File.Exists(TestPath)) File.Delete(TestPath);
     }
 
     public void Dispose()
     {
-        if (File.Exists(TestPath))
-        {
-            File.Delete(TestPath);
-        }
+        if (File.Exists(TestPath)) File.Delete(TestPath);
     }
 
     [Fact]
@@ -48,6 +41,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(SByte.MaxValue);
@@ -77,6 +71,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(UInt16.MaxValue);
@@ -104,6 +99,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         smallest.AddRange(range);
         var check1 = smallest[2];
         check1.Should().Be(2);
@@ -131,6 +127,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(Int64.MaxValue);
@@ -157,6 +154,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(0);
         smallest.SetLast(0);
@@ -183,6 +181,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(Byte.MaxValue);
@@ -209,6 +208,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(UInt16.MaxValue);
@@ -235,6 +235,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(UInt24AsInt64.MaxValue);
@@ -262,6 +263,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(Int24AsInt64.MaxValue);
@@ -291,6 +293,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(UInt32.MaxValue);
@@ -317,6 +320,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(UInt40AsInt64.MaxValue);
@@ -343,6 +347,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(UInt48AsInt64.MaxValue);
@@ -369,6 +374,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(UInt56AsInt64.MaxValue);
@@ -395,6 +401,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(Int40AsInt64.MaxValue);
@@ -424,6 +431,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(Int48AsInt64.MaxValue);
@@ -453,6 +461,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(Int56AsInt64.MaxValue);
@@ -483,6 +492,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(2);
         smallest.SetLast(Int64.MaxValue);
@@ -509,6 +519,7 @@ public class SmallestInt64Tests : IDisposable
         {
             // Check that enumeration works
         }
+
         var check1 = smallest[2];
         check1.Should().Be(0);
         smallest.SetLast(0);
@@ -517,79 +528,5 @@ public class SmallestInt64Tests : IDisposable
         smallest.SetLast(0);
         var checkLast2 = smallest[smallest.Count - 1];
         checkLast2.Should().Be(0);
-    }
-
-    [Fact]
-    public void GetHeaderInfo_Test()
-    {
-        using var smallest = new SmallestInt64ListMmf(DataType.SByte, TestPath);
-        smallest.DataType.Should().Be(DataType.SByte);
-        smallest.WidthBits.Should().Be(8, "sbyte");
-        smallest.Add(1);
-        var (version, dataType, count) = UtilsListMmf.GetHeaderInfo(TestPath);
-        version.Should().Be(0);
-        dataType.Should().Be(DataType.SByte);
-        count.Should().Be(1);
-    }
-
-    [Fact]
-    public void GetHeaderInfoBitArray_Test()
-    {
-        using var smallest = new SmallestInt64ListMmf(DataType.Bit, TestPath);
-        smallest.DataType.Should().Be(DataType.Bit);
-        smallest.WidthBits.Should().Be(1, "BitArray");
-        smallest.Add(1);
-        var (version, dataType, count) = UtilsListMmf.GetHeaderInfo(TestPath);
-        version.Should().Be(0);
-        dataType.Should().Be(DataType.Bit);
-        count.Should().Be(1);
-    }
-
-    [Fact]
-    public void UpgradeSByteToInt32_Test()
-    {
-        using var list = new SmallestInt64ListMmf(DataType.SByte, TestPath);
-        list.DataType.Should().Be(DataType.SByte);
-        list.WidthBits.Should().Be(8, "sbyte");
-        list.Add(1);
-
-        var (_, dataType, _) = UtilsListMmf.GetHeaderInfo(TestPath);
-        dataType.Should().Be(DataType.SByte);
-
-        list.Add(sbyte.MaxValue + 1);
-        var (_, dataType2, _) = UtilsListMmf.GetHeaderInfo(TestPath);
-        dataType2.Should().Be(DataType.Byte);
-    }
-
-    [Fact]
-    public void UpgradeBitArrayToInt32_Test()
-    {
-        using var list = new SmallestInt64ListMmf(DataType.Bit, TestPath);
-        list.DataType.Should().Be(DataType.Bit);
-        list.WidthBits.Should().Be(1, "bit");
-        list.Add(1);
-
-        var (_, dataType, _) = UtilsListMmf.GetHeaderInfo(TestPath);
-        dataType.Should().Be(DataType.Bit);
-
-        list.Add(10);
-        var (_, dataType2, _) = UtilsListMmf.GetHeaderInfo(TestPath);
-        dataType2.Should().Be(DataType.Byte);
-    }
-
-    [Fact]
-    public void AutoUpgradeInt16ToInt32_Test()
-    {
-        // ReSharper disable once UseObjectOrCollectionInitializer
-        using var smallest = new SmallestInt64ListMmf(DataType.Int16, TestPath);
-        smallest.Add(52);
-        var path = smallest.Path;
-        File.Exists(path).Should().BeTrue();
-        smallest.DataType.Should().Be(DataType.Int16);
-        smallest.Add(int.MinValue);
-        smallest.DataType.Should().Be(DataType.Int32);
-        smallest.Count().Should().Be(2);
-        smallest[0].Should().Be(52);
-        smallest[1].Should().Be(int.MinValue);
     }
 }
